Simplify cargo features by making `x86_64` a target-specific dependency
This enables the `MmioSerialPort` type unconditionally on all architectures. The `SerialPort` type is only enabled on `x86_64`.
