+++ src/x86_64.rs
@@ -1,0 +1,119 @@
+use x86_64::instructions::port::Port;
+use crate::LineStsFlags;
+use core::fmt;
+
+/// An interface to a serial port that allows sending out individual bytes.
+pub struct SerialPort {
+    data: Port<u8>,
+    int_en: Port<u8>,
+    fifo_ctrl: Port<u8>,
+    line_ctrl: Port<u8>,
+    modem_ctrl: Port<u8>,
+    line_sts: Port<u8>,
+}
+
+impl SerialPort {
+    /// Creates a new serial port interface on the given I/O port.
+    ///
+    /// This function is unsafe because the caller must ensure that the given base address
+    /// really points to a serial port device.
+    #[cfg(feature = "nightly")]
+    pub const unsafe fn new(base: u16) -> Self {
+        Self {
+            data: Port::new(base),
+            int_en: Port::new(base + 1),
+            fifo_ctrl: Port::new(base + 2),
+            line_ctrl: Port::new(base + 3),
+            modem_ctrl: Port::new(base + 4),
+            line_sts: Port::new(base + 5),
+        }
+    }
+
+    /// Creates a new serial port interface on the given I/O port.
+    ///
+    /// This function is unsafe because the caller must ensure that the given base address
+    /// really points to a serial port device.
+    #[cfg(feature = "stable")]
+    pub unsafe fn new(base: u16) -> Self {
+        Self {
+            data: Port::new(base),
+            int_en: Port::new(base + 1),
+            fifo_ctrl: Port::new(base + 2),
+            line_ctrl: Port::new(base + 3),
+            modem_ctrl: Port::new(base + 4),
+            line_sts: Port::new(base + 5),
+        }
+    }
+
+    /// Initializes the serial port.
+    ///
+    /// The default configuration of [38400/8-N-1](https://en.wikipedia.org/wiki/8-N-1) is used.
+    pub fn init(&mut self) {
+        unsafe {
+            // Disable interrupts
+            self.int_en.write(0x00);
+
+            // Enable DLAB
+            self.line_ctrl.write(0x80);
+
+            // Set maximum speed to 38400 bps by configuring DLL and DLM
+            self.data.write(0x03);
+            self.int_en.write(0x00);
+
+            // Disable DLAB and set data word length to 8 bits
+            self.line_ctrl.write(0x03);
+
+            // Enable FIFO, clear TX/RX queues and
+            // set interrupt watermark at 14 bytes
+            self.fifo_ctrl.write(0xC7);
+
+            // Mark data terminal ready, signal request to send
+            // and enable auxilliary output #2 (used as interrupt line for CPU)
+            self.modem_ctrl.write(0x0B);
+
+            // Enable interrupts
+            self.int_en.write(0x01);
+        }
+    }
+
+    fn line_sts(&mut self) -> LineStsFlags {
+        unsafe { LineStsFlags::from_bits_truncate(self.line_sts.read()) }
+    }
+
+    /// Sends a byte on the serial port.
+    pub fn send(&mut self, data: u8) {
+        unsafe {
+            match data {
+                8 | 0x7F => {
+                    wait_for!(self.line_sts().contains(LineStsFlags::OUTPUT_EMPTY));
+                    self.data.write(8);
+                    wait_for!(self.line_sts().contains(LineStsFlags::OUTPUT_EMPTY));
+                    self.data.write(b' ');
+                    wait_for!(self.line_sts().contains(LineStsFlags::OUTPUT_EMPTY));
+                    self.data.write(8)
+                },
+                _ => {
+                    wait_for!(self.line_sts().contains(LineStsFlags::OUTPUT_EMPTY));
+                    self.data.write(data);
+                },
+            }
+        }
+    }
+
+    /// Receives a byte on the serial port.
+    pub fn receive(&mut self) -> u8 {
+        unsafe {
+            wait_for!(self.line_sts().contains(LineStsFlags::INPUT_FULL));
+            self.data.read()
+        }
+    }
+}
+
+impl fmt::Write for SerialPort {
+    fn write_str(&mut self, s: &str) -> fmt::Result {
+        for byte in s.bytes() {
+            self.send(byte);
+        }
+        Ok(())
+    }
+}
